requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
multi-key-dict==2.0.3
pbr==6.0.0
python-jenkins==1.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.7
